backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
google-generativeai
fish_audio_sdk